Code is now able to deal with !DOCTYPE in XML documents
Although this may not be what you'd expect. Found OPF package file XML documents with improper DOCTYPE declarations (one was claiming the document was html). This led to a parse failure that I was unable to correct for with hxt options. This repair is an ugly hack to remove any <!DOCTYPE ...> declarations encountered before XML parse.