Began large redesign for supporting EPUB3
Much work done to split up various parts of parsing EPUB meta
information into smaller steps. The goal here is to provide something
like an eDSL to allow consumers to do only the work they need but
mix-and-match what parts of an EPUB's meta information they want.

This redesign is still in-progress.